brain_mri_seg_vnet/pipeline/requirements.txt
albumentations
click
google-cloud-aiplatform
google-cloud-storage
monai
munch
natsort
opencv-python
pathlib
Pillow
torch-model-archiver
torchserve
torch-summary
torchvision
tensorboard
scikit-image
scikit-learn





